go test -test.bench=BenchmarkOptimize -test.run=^XXX

```
pkg: github.com/pingcap/tidb/plan
BenchmarkOptimize/select_count(*)_from_t_group_by_e-4         	   30000	     39279 ns/op	   27376 B/op	     370 allocs/op
BenchmarkOptimize/select_count(*)_from_t_where_e_<=_10_group_by_e-4         	   10000	    100107 ns/op	   67394 B/op	     855 allocs/op
BenchmarkOptimize/select_count(*)_from_t_where_e_<=_50-4                    	   10000	    102730 ns/op	   68466 B/op	     861 allocs/op
BenchmarkOptimize/select_count(*)_from_t_where_c_>_'1'_group_by_b-4         	   10000	    113519 ns/op	   78067 B/op	    1079 allocs/op
BenchmarkOptimize/select_count(*)_from_t_where_e_=_1_group_by_b-4           	   10000	    116716 ns/op	   78898 B/op	    1036 allocs/op
BenchmarkOptimize/select_count(*)_from_t_where_e_>_1_group_by_b-4           	   10000	    113030 ns/op	   81554 B/op	    1052 allocs/op
BenchmarkOptimize/select_count(e)_from_t_where_t.b_<=_20-4                  	   10000	    125467 ns/op	   84153 B/op	    1064 allocs/op
BenchmarkOptimize/select_count(e)_from_t_where_t.b_<=_30-4                  	   10000	    138241 ns/op	   84155 B/op	    1065 allocs/op
BenchmarkOptimize/select_count(e)_from_t_where_t.b_<=_40-4                  	   10000	    134946 ns/op	   84155 B/op	    1065 allocs/op
BenchmarkOptimize/select_count(e)_from_t_where_t.b_<=_50-4                  	   10000	    122163 ns/op	   84154 B/op	    1065 allocs/op
BenchmarkOptimize/select_*_from_t_where_t.b_<=_40-4                         	   30000	     55954 ns/op	   41520 B/op	     502 allocs/op
BenchmarkOptimize/select_*_from_t_where_t.b_<=_50-4                         	   30000	     59545 ns/op	   41520 B/op	     502 allocs/op
BenchmarkOptimize/select_*_from_t_where_1_and_t.b_<=_50-4                   	   30000	     55696 ns/op	   41824 B/op	     508 allocs/op
BenchmarkOptimize/select_*_from_t_where_t.b_<=_100_order_by_t.a_limit_1-4   	   10000	    210889 ns/op	  156849 B/op	    1841 allocs/op
BenchmarkOptimize/select_*_from_t_where_t.b_<=_1_order_by_t.a_limit_10-4    	   10000	    213800 ns/op	  155407 B/op	    1804 allocs/op
BenchmarkOptimize/select_*_from_t_use_index(b)_where_b_=_1_order_by_a-4     	   50000	     38628 ns/op	   28159 B/op	     371 allocs/op
BenchmarkOptimize/select_*_from_t_where_d_<_cast('1991-09-05'_as_datetime)-4         	   20000	     64924 ns/op	   41585 B/op	     529 allocs/op
BenchmarkOptimize/select_*_from_t_where_ts_<_'1991-09-05'-4                          	   20000	     73385 ns/op	   41682 B/op	     533 allocs/op
```