@shenli 2017-12-19
You could add some point queries. Like `select * from t where t.a=1;`